+++ crates/nostr/src/event/tag/mod.rs
@@ -158,12 +158,6 @@ impl Tag {
         }
     }
 
-    /// Get reference of array of strings
-    #[deprecated(since = "0.35.0", note = "Use `as_slice` instead.")]
-    pub fn as_vec(&self) -> &[String] {
-        &self.buf
-    }
-
     /// Get as slice of strings
     #[inline]
     pub fn as_slice(&self) -> &[String] {